Add Explicit Hydrogens
The parameter coverage was previously incorrect because we were not
adding explicit hydrogens after loading the SMILES strings. Now, they
should be correct. The hydrogens were added in the notebook and the
script. The notebook was only rerun for Roche and eMolecules first 1000,
as eMolecules first 10000 would take a while.